Denying consent is better if undetectable as such (#476)
* Denying consent is better if undetectable as such

This is not a full generalization of the concepts that discussed in #470 and #475, but I think that it suffices.

Closes #475.

* Link feature detection to consent

This is based on @dbaron's excellent feedback.  However, I took an extra
step with the last sentence here, which I'm not committed to.  There's
an argument to be had that anything like this probably shouldn't be part
of the web platform.

Now, with the controversy established, go!

* Typu

* can-be

Co-authored-by: Amy Guy <amy@rhiaro.co.uk>

---------

Co-authored-by: Amy Guy <amy@rhiaro.co.uk>