Use names that describe purpose
I found that I needed to redo the whole section here.
Changing the focus seemed better than adding text.
This is more text, but not a lot more.

Future-proofing is one motivation, but it's not the real motivation for
these recommendations.

I also found the key event example to be spectacularly non-compelling.
It even made me think that it is almost a bad example.
Key events can be synthesized by speech recognition tools in some cases.

To that end, I chose an example that I'm more familiar with.

Closes #507.