Clarify the strings for enums/constants principle
Adds examples and tightens things up.

Adds a mention that WebIDL enums only take string values.

One thing that jumped out here was that WebIDL permits the use of
numeric/integer constants.  I searched all of the Firefox codebase and I
didn't find a single useful instance where a numeric constant made
sense.

My operating theory before starting was that there might be some case
where developers might be exposed to a number in contexts outside of the
web, which might mean that using a numeric constant in the platform made
sense.

The one example of that we have is KeyboardEvent.keyCode, which might
have made sense if it exposed the Unicode codepoint of a key.  That,
however, is solidly deprecated in favor of a much better API.

Every other instance was a bad example.

Closes #427.